Fixed strlcpy

+++ ft_strlen.c
@@ -1,6 +1,8 @@
-unsigned long int ft_strlen(const char *s)
+#include <stdio.h>
+
+size_t ft_strlen(const char *s)
 {
-  unsigned long int i;
+  size_t i;
 
   i = 0;
   while (s[i])
